LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-spline-invalid-keyTimes-expected.txt


FAIL keyTimes does not end with one, calcMode=spline assert_equals: motion path not applied expected 0 but got -10
